requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via discord-py
aiosignal==1.3.2
    # via aiohttp
attrs==24.3.0
    # via aiohttp
black==24.10.0
    # via -r requirements.in
click==8.1.7
    # via black
discord==2.3.2
    # via -r requirements.in
discord-py==2.4.0
    # via discord
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
load-dotenv==0.1.0
    # via -r requirements.in
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy==1.13.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==24.2
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
python-dotenv==1.0.1
    # via load-dotenv
ruff==0.8.4
    # via -r requirements.in
typing-extensions==4.12.2
    # via mypy
yarl==1.18.3
    # via aiohttp
